+++ demo/js/hidden-treasures.js
@@ -1,4 +1,4 @@
-var { focusTrap: createFocusTrap } = require('../../dist/focus-trap');
+var { createFocusTrap } = require('../../dist/focus-trap');
 
 var container = document.getElementById('ht');
 var more = document.getElementById('ht-more');